Moved old core server code to an example file.
